+++ src/include/main/db_main.h
@@ -17,8 +17,8 @@
 #include "network/postgres/postgres_protocol_interpreter.h"
 #include "optimizer/statistics/stats_storage.h"
 #include "self_driving/model_server/model_server_manager.h"
-#include "self_driving/pilot/pilot.h"
-#include "self_driving/pilot/pilot_thread.h"
+#include "self_driving/planning/pilot.h"
+#include "self_driving/planning/pilot_thread.h"
 #include "settings/settings_manager.h"
 #include "settings/settings_param.h"
 #include "storage/garbage_collector_thread.h"
