--- src/self_driving/pilot/action/abstract_action.cpp
@@ -1,5 +1,0 @@
-#include "self_driving/pilot/action/abstract_action.h"
-
-namespace noisepage::selfdriving::pilot {
-action_id_t AbstractAction::action_id_counter{0};
-}  // namespace noisepage::selfdriving::pilot
