+++ src/self_driving/planning/action/abstract_action.cpp
@@ -1,0 +1,5 @@
+#include "self_driving/planning/action/abstract_action.h"
+
+namespace noisepage::selfdriving::pilot {
+action_id_t AbstractAction::action_id_counter{0};
+}  // namespace noisepage::selfdriving::pilot
